requirements.txt
numpy==1.18.4
matplotlib==3.1.3
pandas==1.0.3
opencv-python==4.2.0.34
pillow==7.1.2
pytorch==1.5.0
scikit-learn==0.23.1
torchvision==0.6.0
